feat: created download button in the bubble component
